Utils corrections && table name changes && fake migrations
