+++ dataworkspace/dataworkspace/apps/applications/models.py
@@ -40,6 +40,7 @@ class ApplicationTemplate(TimeStampedModel):
     )
 
     class Meta:
+        db_table = "app_applicationtemplate"
         indexes = [
             models.Index(fields=['name']),
         ]
@@ -103,6 +104,7 @@ class ApplicationInstance(TimeStampedModel):
     )
 
     class Meta:
+        db_table = "app_applicationinstance"
         indexes = [
             models.Index(fields=['owner', 'created_date']),
             models.Index(fields=['public_host', 'state']),
