+++ dataworkspace/dataworkspace/apps/applications/utils.py
@@ -1,7 +1,17 @@
 import hashlib
+import logging
+import requests
+import urllib.parse
+
+from datetime import datetime
+
+from django.conf import settings
 
 from dataworkspace.apps.applications.spawner import get_spawner
 from dataworkspace.apps.applications.models import ApplicationInstance
+from dataworkspace.cel import celery_app
+
+logger = logging.getLogger('app')
 
 
 def api_application_dict(application_instance):
@@ -18,7 +28,6 @@ def api_application_dict(application_instance):
     api_state = \
         application_instance.state if spawner_state == 'RUNNING' else \
         spawner_state
-
 
     template_name = application_instance.application_template.name
     sso_id_hex = hashlib.sha256(str(application_instance.owner.profile.sso_id).encode('utf-8')).hexdigest()
@@ -50,3 +59,92 @@ def application_api_is_allowed(request, public_host):
 
     return (owner_sso_id_hex == request_sso_id_hex[:8] and
             request.user.has_perm('applications.start_all_applications'))
+
+
+def stop_spawner_and_application(application_instance):
+    get_spawner(application_instance.spawner).stop(
+        application_instance.spawner_application_template_options,
+        application_instance.spawner_application_instance_id,
+    )
+    set_application_stopped(application_instance)
+
+
+def set_application_stopped(application_instance):
+    application_instance.state = 'STOPPED'
+    application_instance.single_running_or_spawning_integrity = str(application_instance.id)
+    application_instance.save()
+
+
+def application_instance_max_cpu(application_instance):
+    # If we don't have the proxy url yet, we can't have any metrics yet.
+    # This is expected and should not be shown as an error
+    if application_instance.proxy_url is None:
+        raise ValueError('Unknown')
+
+    instance = urllib.parse.urlsplit(application_instance.proxy_url).hostname + ':8889'
+    url = f'https://{settings.PROMETHEUS_DOMAIN}/api/v1/query'
+    params = {
+        'query': f'increase(precpu_stats__cpu_usage__total_usage{{instance="{instance}"}}[30s])[2h:30s]'
+    }
+    try:
+        response = requests.get(url, params)
+    except requests.RequestException:
+        raise ValueError('Error connecting to metrics server')
+
+    response_dict = response.json()
+    if response_dict['status'] != 'success':
+        raise ValueError(f'Metrics server return value is {response_dict["status"]}')
+
+    try:
+        values = response_dict['data']['result'][0]['values']
+    except (IndexError, KeyError):
+        # The server not having metrics yet should not be reported as an error
+        raise ValueError(f'Unknown')
+
+    max_cpu = 0.0
+    ts_at_max = 0
+    for ts, cpu in values:
+        cpu_float = float(cpu) / (1000000000 * 30) * 100
+        if cpu_float >= max_cpu:
+            max_cpu = cpu_float
+            ts_at_max = ts
+
+    return max_cpu, ts_at_max
+
+
+@celery_app.task()
+def kill_idle_fargate():
+    logger.info('kill_idle_fargate: Start')
+
+    two_hours_ago = datetime.datetime.now(datetime.timezone.utc) + datetime.timedelta(hours=-2)
+    instances = ApplicationInstance.objects.filter(
+        spawner='FARGATE',
+        state__in=['RUNNING', 'SPAWNING'],
+        created_date__lt=two_hours_ago,
+    )
+
+    for instance in instances:
+        if instance.state == 'SPAWNING':
+            set_application_stopped(instance)
+            continue
+
+        logger.info('kill_idle_fargate: Attempting to find CPU usage of %s', instance)
+        try:
+            max_cpu, _ = application_instance_max_cpu(instance)
+        except Exception:
+            logger.exception('kill_idle_fargate: Unable to find CPU usage for %s', instance)
+            continue
+
+        logger.info('kill_idle_fargate: CPU usage for %s is %s', instance, max_cpu)
+
+        if max_cpu >= 1.0:
+            continue
+
+        try:
+            stop_spawner_and_application(instance)
+        except Exception:
+            logger.exception('kill_idle_fargate: Unable to stop %s', instance)
+
+        logger.info('kill_idle_fargate: Stopped application %s', instance)
+
+    logger.info('kill_idle_fargate: End')
