+++ dataworkspace/dataworkspace/apps/dw_admin/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class DWAdminConfig(AppConfig):
+    name = 'dw_admin'
