+++ dataworkspace/dataworkspace/settings/base.py
@@ -165,7 +165,7 @@ ZENDESK_SERVICE_FIELD_VALUE = env['ZENDESK_SERVICE_FIELD_VALUE']
 CELERY_BROKER_URL = env['REDIS_URL']
 CELERY_BEAT_SCHEDULE = {
     'kill-idle-fargate-containers': {
-        'task': 'dataworkspace.apps.core.utils.kill_idle_fargate',
+        'task': 'dataworkspace.apps.applications.utils.kill_idle_fargate',
         'schedule': 60 * 10,
         'args': (),
     },
